+++ src/commands/Folders.ts
@@ -219,7 +219,9 @@ export class Folders {
           : Folders.getAbsFilePath(assetFolder);
       const wsFolder = Folders.getWorkspaceFolder();
       if (wsFolder) {
-        const relativePath = relative(parseWinPath(wsFolder.fsPath), parseWinPath(assetFolder));
+        const relativePath = parseWinPath(
+          relative(parseWinPath(wsFolder.fsPath), parseWinPath(assetFolder))
+        );
         return relativePath === '' ? '/' : relativePath;
       }
     }
@@ -648,7 +650,7 @@ export class Folders {
 
     const uniqueFolders = [...new Set(folders)];
     const relativeFolderPaths = uniqueFolders.map((folder) =>
-      relative(parseWinPath(wsFolder.fsPath), folder)
+      parseWinPath(relative(parseWinPath(wsFolder.fsPath), folder))
     );
 
     Logger.verbose('Folders:getContentFolders:end');